talentum-backend/requirements.txt
fastapi
uvicorn
pymongo
neo4j
python-decouple
pydantic
redis
networkx